# docs/swagger.yaml
definitions:
  request.ChallengeCreateRequest:
    properties:
      category:
        default: misc
        type: string
      description:
        default: 题目描述
        type: string
      difficulty:
        default: 1
        type: integer
      duration:
        default: 1800
        type: integer
      exposed_port:
        default: 80
        type: integer
      flag:
        default: PgsCTF{Th4nk5_4_us1ng_PgsHub}
        type: string
      flag_env:
        default: FLAG
        type: string
      flag_fmt:
        default: PgsCTF{[UUID]}
        type: string
      has_attachment:
        default: false
        type: boolean
      image:
        default: nginx
        type: string
      is_dynamic:
        default: true
        type: boolean
      is_practicable:
        default: false
        type: boolean
      memory_limit:
        default: 512
        type: integer
      practice_pts:
        default: 200
        type: integer
      title:
        default: 新题目
        type: string
    type: object
  request.ChallengeDeleteRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  request.ChallengeUpdateRequest:
    properties:
      category:
        type: string
      description:
        type: string
      difficulty:
        type: integer
      duration:
        type: integer
      exposed_port:
        type: integer
      flag:
        type: string
      flag_env:
        type: string
      flag_fmt:
        type: string
      has_attachment:
        type: boolean
      id:
        type: string
      image:
        type: string
      is_dynamic:
        type: boolean
      is_practicable:
        type: boolean
      memory_limit:
        type: integer
      practice_pts:
        type: integer
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: 没有其他东西啦，仅仅是所有的后端接口，不要乱用哦
  title: PgsHub Backend API
  version: "1.0"
paths:
  /api/assets/games/cover/{id}:
    get:
      consumes:
      - application/json
      description: 通过比赛 Id 获取比赛封面
      parameters:
      - description: 比赛 Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 通过比赛 Id 获取比赛封面
      tags:
      - 资源
    post:
      consumes:
      - multipart/form-data
      description: 通过比赛 Id 设置比赛封面
      parameters:
      - description: 比赛 Id
        in: path
        name: id
        required: true
        type: string
      - description: 封面文件
        in: formData
        name: avatar
        required: true
        type: file
      responses: {}
      summary: 通过比赛 Id 设置比赛封面
      tags:
      - 资源
  /api/assets/games/writeups/{id}:
    get:
      consumes:
      - application/json
      description: 通过团队 Id 获取比赛 Writeup
      parameters:
      - description: 团队 Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 通过团队 Id 获取比赛 Writeup
      tags:
      - 资源
...
  /api/assets/teams/avatar/{id}:
    get:
      consumes:
      - application/json
      description: 通过团队 Id 获取团队头像
      parameters:
      - description: 团队 Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 通过团队 Id 获取团队头像
      tags:
      - 资源
    post:
      consumes:
      - multipart/form-data
      description: 通过团队 Id 设置团队头像
      parameters:
      - description: 团队 Id
        in: path
        name: id
        required: true
        type: string
      - description: 头像文件
        in: formData
        name: avatar
        required: true
        type: file
      responses: {}
      summary: 通过团队 Id 设置团队头像
      tags:
      - 资源
  /api/assets/teams/avatar/{id}/exists:
    get:
      consumes:
      - application/json
      description: 通过团队 Id 确认团队头像是否存在
      parameters:
      - description: 团队 Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 通过团队 Id 确认团队头像是否存在
      tags:
      - 资源
...
  /api/assets/users/avatar/{id}:
    delete:
      consumes:
      - application/json
      description: 通过用户 Id 删除用户头像
      parameters:
      - description: 用户 Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 通过用户 Id 删除用户头像
      tags:
      - 资源
    get:
      consumes:
      - application/json
      description: 通过用户 Id 获取用户头像
      parameters:
      - description: 用户 Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 通过用户 Id 获取用户头像
      tags:
      - 资源
    post:
      consumes:
      - multipart/form-data
      description: 通过用户 Id 设置用户头像
      parameters:
      - description: 用户 Id
        in: path
        name: id
        required: true
        type: string
      - description: 头像文件
        in: formData
        name: avatar
        required: true
        type: file
      responses: {}
      summary: 通过用户 Id 设置用户头像
      tags:
      - 资源
  /api/assets/users/avatar/{id}/exists:
    get:
      consumes:
      - application/json
      description: 通过用户 Id 确认用户头像是否存在
      parameters:
      - description: 用户 Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 通过用户 Id 确认用户头像是否存在
      tags:
      - 资源
  /api/challenges/:
    delete:
      consumes:
      - application/json
      parameters:
      - description: PgsToken
        in: header
        name: PgsToken
        required: true
        type: string
      - description: ChallengeDeleteRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ChallengeDeleteRequest'
      produces:
      - application/json
      responses: {}
      summary: 删除题目（Role≤2）
      tags:
      - 题目
    get:
      consumes:
      - application/json
      description: 只有当 Role≤2 并且 IsDetailed=1 时，才会提供题目的关键信息
      parameters:
      - description: PgsToken
        in: header
        name: PgsToken
        required: true
        type: string
      - in: query
        name: category
        type: string
      - in: query
        name: difficulty
        type: integer
      - in: query
        name: id
        type: string
      - in: query
        name: is_detailed
        type: integer
      - in: query
        name: is_dynamic
        type: integer
      - in: query
        name: is_practicable
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: size
        type: integer
      - in: query
        name: title
        type: string
      produces:
      - application/json
      responses: {}
      summary: 题目查询
      tags:
      - 题目
    post:
      consumes:
      - application/json
      parameters:
      - description: PgsToken
        in: header
        name: PgsToken
        required: true
        type: string
      - description: ChallengeCreateRequest
        in: body
        name: 创建请求
        required: true
        schema:
          $ref: '#/definitions/request.ChallengeCreateRequest'
      produces:
      - application/json
      responses: {}
      summary: 创建题目（Role≤2）
      tags:
      - 题目
    put:
      consumes:
      - application/json
      parameters:
      - description: PgsToken
        in: header
        name: PgsToken
        required: true
        type: string
      - description: ChallengeUpdateRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ChallengeUpdateRequest'
      produces:
      - application/json
      responses: {}
      summary: 更新题目（Role≤2）
      tags:
      - 题目
...
swagger: "2.0"